backoffice-api: [Mongo DB working fine with Spring Boot: added Mongo setup instructions. Campaigns collection primary key type was changed from Integer to String]
