+++ backoffice-api/src/main/java/com/tipico/social/marketing/controller/CampaignsController.java
@@ -1,7 +1,8 @@
 package com.tipico.social.marketing.controller;
 
-import com.tipico.social.marketing.entity.Campaign;
+import com.tipico.social.marketing.contract.Campaign;
 import com.tipico.social.marketing.repository.CampaignsRepository;
+import com.tipico.social.marketing.service.CampaignsService;
 import org.codehaus.jackson.map.ObjectMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
@@ -24,7 +25,7 @@ import java.util.List;
 public class CampaignsController {
 
     @Autowired
-    private CampaignsRepository campaignsRepository;
+    private CampaignsService campaignsService;
 
     @RequestMapping("/")
     public String root() {
@@ -34,72 +35,38 @@ public class CampaignsController {
     @RequestMapping("/getCampaign")
     @ResponseBody
     public Campaign getCampaign(@RequestParam(value="id", required=true) String id) {
-        try {
-            return campaignsRepository.findOne(Integer.parseInt(id));
-        } catch (Throwable t) {
-            t.printStackTrace();
-        }
-        return null;
+        return campaignsService.getCampaign(id);
     }
 
     @RequestMapping("/saveCampaign")
     @ResponseBody
     public Campaign saveCampaign(@RequestParam(value="campaign", required=true) String campaignJson) {
-        try {
-            Campaign campaign = new ObjectMapper().readValue(campaignJson, Campaign.class);
-            return campaignsRepository.insert(campaign);
-        } catch (Throwable t) {
-            t.printStackTrace();
-        }
-        return null;
+        return campaignsService.saveCampaign(campaignJson);
     }
 
     @RequestMapping("/deleteCampaign")
     @ResponseBody
     public void deleteCampaign(@RequestParam(value="id", required=true) String id) {
-        try {
-            Campaign campaign = campaignsRepository.findOne(Integer.parseInt(id));
-            campaignsRepository.delete(campaign);
-        } catch (Throwable t) {
-            t.printStackTrace();
-        }
+        campaignsService.deleteCampaign(id);
     }
 
     @RequestMapping("/updateCampaign")
     @ResponseBody
     public Campaign updateCampaign(@RequestParam(value="campaign", required=true) String campaignJson) {
-        try {
-            Campaign campaign = new ObjectMapper().readValue(campaignJson, Campaign.class);
-            return campaignsRepository.save(campaign);
-        } catch (Throwable t) {
-            t.printStackTrace();
-        }
-        return null;
+        return campaignsService.updateCampaign(campaignJson);
     }
 
-    @RequestMapping("/countCampaign")
+    @RequestMapping("/countCampaigns")
     @ResponseBody
-    public long countCampaign() {
-        try {
-            return campaignsRepository.count();
-        } catch (Throwable t) {
-            t.printStackTrace();
-        }
-        return -1L;
+    public long countCampaigns() {
+        return campaignsService.countCampaigns();
     }
 
     @RequestMapping("/requestPagedCampaigns")
     @ResponseBody
     public List<Campaign> requestPagedCampaigns(@RequestParam(value="pageNumber", required=true) int pageNumber,
                                                 @RequestParam(value="campaignsPerPage", required=true) int perPage) {
-        List<Campaign> campaignList = null;
-
-        try {
-            campaignList = campaignsRepository.findAll(new PageRequest(pageNumber, perPage)).getContent();
-        } catch (Throwable t) {
-            t.printStackTrace();
-        }
-        return campaignList;
+        return campaignsService.requestPagedCampaigns(pageNumber, perPage);
     }
 
     public static void main(String[] args) throws Exception {
