+++ backoffice-api/src/main/java/com/tipico/social/marketing/repository/CampaignsRepository.java
@@ -1,6 +1,6 @@
 package com.tipico.social.marketing.repository;
 
-import com.tipico.social.marketing.entity.Campaign;
+import com.tipico.social.marketing.contract.Campaign;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.data.repository.RepositoryDefinition;
 
@@ -8,5 +8,4 @@ import org.springframework.data.repository.RepositoryDefinition;
  * Created by ncamilleri on 16/04/16.
  */
 @RepositoryDefinition(domainClass = Campaign.class, idClass = Integer.class)
-public interface CampaignsRepository extends MongoRepository<Campaign, Integer> {
-}
+public interface CampaignsRepository extends MongoRepository<Campaign, Integer> {}
