+++ contract/src/main/java/com/tipico/social/marketing/contract/Campaign.java
@@ -1,10 +1,12 @@
 package com.tipico.social.marketing.contract;
 
 import java.util.Date;
+import org.springframework.data.annotation.Id;
 
 public class Campaign {
 
-	private Integer id;
+	@Id
+	private String id;
 	private String eventName;
 	private Date created;
 	private Date lastUpdated;
@@ -13,11 +15,11 @@ public class Campaign {
 	private int delayBetweenPosts;
 	private String message;
 
-	public Integer getId() {
+	public String getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(String id) {
 		this.id = id;
 	}
